refactor: Separate backend and frontend, plus add a main entrypoint.
