+++ src/__main__.py
@@ -1,0 +1,5 @@
+from src.frontend.main_window import TicketCounterApp
+
+if __name__ == "__main__":
+    app = TicketCounterApp()
+    app.run()
