+++ src/frontend/add_denomination_window.py
